requirements.txt
pillow
pandas
bs4
googletrans
deep_translator
korean-romanizer
